Fix handling of cached errors
PageViewInfoService::getPageData() should return a fatal status
iff data loading for every single page failed. (This is probably
not a good idea but beyond the scope of this patch.) The case
when 0 titles were requested is special-cased (as it wouldn't
make sense to treat that as an error), but it was done
incorrectly so the case where all pages fail with cached errors
was also treated as non-fatal.

Bug: T181012
Change-Id: I5daf02efb776076def1d303bd02cace3015fbcdd